Pulled TearOff to Shell
